add a few extra warning flags
- in retrospect some of these were probably overkill. Still I left them there in the form of -Wno-error so they don't halt the build..

+++ src/main/drivers/accgyro_mpu.c
@@ -156,7 +156,7 @@ static void mpu6050FindRevision(void)
 
     // determine product ID and accel revision
     ack = mpuConfiguration.read(MPU_RA_XA_OFFS_H, 6, readBuffer);
-    revision = ((readBuffer[5] & 0x01) << 2) | ((readBuffer[3] & 0x01) << 1) | (readBuffer[1] & 0x01);
+    revision = (uint8_t)(((readBuffer[5] & 0x01) << 2) | ((readBuffer[3] & 0x01) << 1) | (readBuffer[1] & 0x01));
     if (revision) {
         /* Congrats, these parts are better. */
         if (revision == 1) {
